pyproject.toml
[tool.poetry]
name = "softdesk-api"
version = "0.1.0"
description = "SoftDesk API REST with Django"
authors = ["degregor69 <raphael.degregorio@gmail.com>"]
readme = "README.md"
packages = [{include = "softdesk_api"}]

[tool.poetry.dependencies]
python = "^3.13.3"
django = "^5.2"
djangorestframework = "^3.16.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
